refactoring HttpClientConfiguration #20
