+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/HttpClientFactory.java
@@ -1,11 +1,15 @@
 package io.github.clescot.kafka.connect.http.client;
 
 import com.google.common.base.Preconditions;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
+import io.github.clescot.kafka.connect.MapUtils;
 import io.github.clescot.kafka.connect.http.client.config.AddSuccessStatusToHttpExchangeFunction;
 import io.github.clescot.kafka.connect.http.client.proxy.ProxySelectorFactory;
 import io.github.clescot.kafka.connect.http.client.ssl.AlwaysTrustManagerFactory;
 import io.micrometer.core.instrument.composite.CompositeMeterRegistry;
 import jakarta.annotation.Nullable;
+import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -24,9 +28,7 @@ import java.net.SocketAddress;
 import java.nio.file.Path;
 import java.security.*;
 import java.security.cert.CertificateException;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Random;
+import java.util.*;
 import java.util.concurrent.ExecutorService;
 import java.util.regex.Pattern;
 
@@ -36,12 +38,13 @@ import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.SUC
 
 /**
  * Factory to build a HttpClient.
+ *
  * @param <C> client type, which is a subclass of HttpClient
  * @param <R> native HttpRequest
  * @param <S> native HttpResponse
  */
-public interface HttpClientFactory<C extends HttpClient<R,S>,R,S> {
-
+public interface HttpClientFactory<C extends HttpClient<R, S>, R, S> {
+    String SHA_1_PRNG = "SHA1PRNG";
     Logger LOGGER = LoggerFactory.getLogger(HttpClientFactory.class);
     String DEFAULT_SSL_PROTOCOL = "SSL";
     String IS_NOT_SET = " is not set";
@@ -108,10 +111,10 @@ public interface HttpClientFactory<C extends HttpClient<R,S>,R,S> {
     }
 
 
-     default KeyManagerFactory getKeyManagerFactory(String keyStorePath,
-                                                          char[] password,
-                                                          @Nullable String keystoreType,
-                                                          @Nullable String algorithm) {
+    default KeyManagerFactory getKeyManagerFactory(String keyStorePath,
+                                                   char[] password,
+                                                   @Nullable String keystoreType,
+                                                   @Nullable String algorithm) {
         KeyManagerFactory keyManagerFactory;
         KeyStore keyStore;
         try {
@@ -127,7 +130,7 @@ public interface HttpClientFactory<C extends HttpClient<R,S>,R,S> {
         File file = path.toFile();
         try (InputStream inputStream = new FileInputStream(file)) {
             keyStore.load(inputStream, password);
-            keyManagerFactory.init(keyStore,password);
+            keyManagerFactory.init(keyStore, password);
         } catch (IOException | NoSuchAlgorithmException | CertificateException | KeyStoreException |
                  UnrecoverableKeyException e) {
             throw new HttpException(e);
@@ -161,11 +164,11 @@ public interface HttpClientFactory<C extends HttpClient<R,S>,R,S> {
         return trustManagerFactory;
     }
 
-    static TrustManagerFactory getTrustManagerFactory(Map<String,String> config){
-        if(config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST)&& Boolean.parseBoolean(config.get(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST).toString())){
+    static TrustManagerFactory getTrustManagerFactory(Map<String, String> config) {
+        if (config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST) && Boolean.parseBoolean(config.get(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST).toString())) {
             LOGGER.warn("/!\\ activating 'always trust any certificate' feature : remote SSL certificates will always be granted. Use this feature at your own risk ! ");
             return new AlwaysTrustManagerFactory();
-        }else {
+        } else {
             String trustStorePath = config.get(HTTP_CLIENT_SSL_TRUSTSTORE_PATH);
             Preconditions.checkNotNull(trustStorePath, CONFIG_HTTP_CLIENT_SSL_TRUSTSTORE_PATH + IS_NOT_SET);
 
@@ -189,31 +192,79 @@ public interface HttpClientFactory<C extends HttpClient<R,S>,R,S> {
 
     static Optional<TrustManagerFactory> buildTrustManagerFactory(Map<String, String> config) {
         if ((config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_PATH)
-                && config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_PASSWORD))||config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST)) {
+                && config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_PASSWORD)) || config.containsKey(HTTP_CLIENT_SSL_TRUSTSTORE_ALWAYS_TRUST)) {
 
             Optional<TrustManagerFactory> trustManagerFactoryOption = Optional.ofNullable(
                     getTrustManagerFactory(config));
             if (trustManagerFactoryOption.isPresent()) {
                 TrustManagerFactory myTrustManagerFactory = trustManagerFactoryOption.get();
-                LOGGER.info("using trustManagerFactory class : {}",myTrustManagerFactory.getClass().getName());
+                LOGGER.info("using trustManagerFactory class : {}", myTrustManagerFactory.getClass().getName());
                 return Optional.of(myTrustManagerFactory);
             }
         }
         return Optional.empty();
     }
+
     static SSLSocketFactory getSSLSocketFactory(@Nullable KeyManagerFactory keyManagerFactory,
-                                                       @Nullable TrustManagerFactory trustManagerFactory,
-                                                       @Nullable String protocol,
-                                                       SecureRandom random){
+                                                @Nullable TrustManagerFactory trustManagerFactory,
+                                                @Nullable String protocol,
+                                                SecureRandom random) {
         try {
             SSLContext sslContext = SSLContext.getInstance(Optional.ofNullable(protocol).orElse(DEFAULT_SSL_PROTOCOL));
-            sslContext.init(keyManagerFactory!=null?keyManagerFactory.getKeyManagers():null,trustManagerFactory!=null?trustManagerFactory.getTrustManagers():null, random);
+            sslContext.init(keyManagerFactory != null ? keyManagerFactory.getKeyManagers() : null, trustManagerFactory != null ? trustManagerFactory.getTrustManagers() : null, random);
             return sslContext.getSocketFactory();
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
             throw new HttpException(e);
         }
     }
 
+
+
+    public static  <C extends HttpClient<R, S>, R, S>Map<String,C> buildConfigurations(
+            HttpClientFactory<C, R, S> httpClientFactory,
+            ExecutorService executorService,
+            List<String> configIdList,
+            Map<String, String> originals, CompositeMeterRegistry meterRegistry) {
+        Map<String,C> httpClientConfigurations = Maps.newHashMap();
+        List<String> configurationIds = Lists.newArrayList();
+        Optional<List<String>> ids = Optional.ofNullable(configIdList);
+        ids.ifPresent(configurationIds::addAll);
+        for (String configId : configurationIds) {
+            Map<String, String> config = Maps.newHashMap(MapUtils.getMapWithPrefix(originals, "config." + configId + "."));
+            HashMap<String, String> settings = Maps.newHashMap(config);
+            settings.put("configuration.id", configId);
+            Random random = getRandom(settings);
+            C httpClient = httpClientFactory.buildHttpClient(settings, executorService, meterRegistry, random);
+            httpClientConfigurations.put(configId,httpClient);
+        }
+        return httpClientConfigurations;
+    }
+
+    @java.lang.SuppressWarnings({"java:S2119","java:S2245"})
+    @NotNull
+    static Random getRandom(Map<String, String> config) {
+        Random random;
+
+        try {
+            if(config.containsKey(HTTP_CLIENT_SECURE_RANDOM_ACTIVATE)&&Boolean.parseBoolean(config.get(HTTP_CLIENT_SECURE_RANDOM_ACTIVATE))){
+                String rngAlgorithm = SHA_1_PRNG;
+                if (config.containsKey(HTTP_CLIENT_SECURE_RANDOM_PRNG_ALGORITHM)) {
+                    rngAlgorithm = config.get(HTTP_CLIENT_SECURE_RANDOM_PRNG_ALGORITHM);
+                }
+                random = SecureRandom.getInstance(rngAlgorithm);
+            }else {
+                if(config.containsKey(HTTP_CLIENT_UNSECURE_RANDOM_SEED)){
+                    long seed = Long.parseLong(config.get(HTTP_CLIENT_UNSECURE_RANDOM_SEED));
+                    random = new Random(seed);
+                }else {
+                    random = new Random();
+                }
+            }
+        } catch (NoSuchAlgorithmException e) {
+            throw new HttpException(e);
+        }
+        return random;
+    }
 
 
 }
